+++ src/modules/components/proposal/proposalActions/actions/generators/proposalAction.ts
@@ -2,7 +2,6 @@ import type { IProposalAction } from '../../proposalActionsTypes';
 
 export const generateProposalAction = (action?: Partial<IProposalAction>): IProposalAction => ({
     type: 'unknownType',
-    contractAddress: '0xContract',
     from: '0xFrom',
     to: '0xTo',
     data: '',