+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/index.ts
@@ -1,3 +1,4 @@
 export type { IProposalAction, IProposalActionInputData, IProposalActionInputDataParameter } from './proposalAction';
-export type { IProposalActionWithdrawToken } from './proposalActionTokenWithdraw';
-export { ProposalActionType } from './ProposalActionType';
+export { ProposalActionComponent, type IProposalActionComponentProps } from './proposalActionComponent';
+export { ProposalActionType } from './proposalActionType';
+export type { IProposalActionWithdrawToken } from './proposalActionWithdrawToken';