+++ src/modules/components/proposal/proposalActions/proposalActionsTypes/proposalAction.ts
@@ -1,56 +1,52 @@
 export interface IProposalActionInputDataParameter {
     /**
-     * The type of the parameter
+     * The type of the parameter (e.g. address, uint256, uint32, ..).
      */
     type: string;
     /**
-     * The value of the parameter
+     * The value of the parameter.
      */
     value: string;
 }
 
 export interface IProposalActionInputData {
     /**
-     * The function to call from proposal action
+     * Name of the function to call from proposal action.
      */
     function: string;
     /**
-     * The contract address to call
+     * The name of the contract to interact with.
      */
     contract: string;
     /**
-     * The parameters to pass to the function
+     * The parameters to pass to the function.
      */
     parameters: IProposalActionInputDataParameter[];
 }
 
 export interface IProposalAction {
     /**
-     * The address of the contract the proposal is attempting to interact with
-     */
-    contractAddress: string;
-    /**
-     * The address to send the transaction from
+     * The address to send the transaction from.
      */
     from: string;
     /**
-     * The address to send the transaction to
+     * The address to send the transaction to.
      */
     to: string;
     /**
-     * The data to send with the transaction
+     * The data to send with the transaction.
      */
     data: string;
     /**
-     * The value to send with the transaction
+     * The value to send with the transaction.
      */
-    value: string;
+    value: string | null;
     /**
-     * The type of the proposal action
+     * The type of the proposal action.
      */
     type: string;
     /**
-     * The input data for the proposal action
+     * The input data for the proposal action.
      */
     inputData: IProposalActionInputData | null;
 }